requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
apturl==0.5.2
async-timeout==4.0.3
attrs==24.2.0
bcrypt==3.2.0
bitarray==2.9.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
ckzg==2.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cytoolz==0.12.3
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
fasteners==0.14.1
frozenlist==1.4.1
future==0.18.2
gyp==0.1
hexbytes==1.2.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
parsimonious==0.10.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.21.0
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyunormalize==16.0.0
pyxdg==0.27
PyYAML==5.4.1
regex==2024.9.11
reportlab==3.6.8
requests==2.25.1
rlp==4.0.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
toolz==0.12.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
wadllib==1.3.6
web3==7.3.0
websockets==13.1
xdg==5
xkit==0.0.0
yarl==1.13.1
zipp==1.0.0
